Move `CompiledBlacklistAwareTrait` under `Checkers/Traits/` namespace
